requirements.txt
numpy
librosa
pyyaml
scipy
musdb